requirements.txt
streamlit>=1.40.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
groq>=0.4.0
PyYAML>=6.0
